splitting code runner from terminal
